requirements.txt
attrs==19.1.0
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
clarifai==2.6.2
configparser==3.7.4
decorator==4.0.11
dropbox==9.4.0
EasyProcess==0.2.7
emoji==0.5.2
future==0.17.1
googleapis-common-protos==1.6.0
grpcio==1.22.0
idna==2.8
imageio==2.1.2
instapy==0.5.5
instapy-chromedriver==2.46
ipython==7.6.0
ipython-genutils==0.2.0
jedi==0.14.0
jsonschema==3.0.1
MeaningCloud-python==1.1.1
moviepy==0.2.3.2
numpy==1.16.4
pandas==0.24.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
plyer==1.4.0
prompt-toolkit==2.0.9
protobuf==3.9.0
ptyprocess==0.6.0
Pygments==2.4.2
pyrsistent==0.15.3
python-dateutil==2.8.0
pytz==2019.1
PyVirtualDisplay==0.2.4
PyYAML==5.1.1
regex==2019.6.8
requests==2.22.0
requests-toolbelt==0.7.0
selenium==3.141.0
six==1.12.0
tqdm==4.11.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
